Add --filterHooks argument
When applied, this patch adds a command line flag (--filterHooks) that
removes every pre/post hooks from the npm scripts menu. This ease action
selection in a project that quite some scripts in it's configuration.

If no arguments or if no arguments matching --filterHooks are present
while invoking nok, the current behavior will remain the same and
everything in the scripts property will be showed in the menu.

If an argument matching --filterHooks is present, a filter function is
applied that remove all script names matching a start with either pre
or post. That should filter pre and post hooks.